Update main.cpp
Created a constant string of the file name that way it is constant among the program and can be easily changed later.

Changed some variable names to match what they are supposed to do and moved them to the top of the functions to make it more readable

Due to changing the file.csv to a constant, there is no longer a need to pass it within the functions

Improved your searching while loop so that it exists once a user is found as well instead of looping through the entire file.